🚚 Rename: 주제별 컴포넌트 폴더로 정리
